pyproject.toml
[tool.poetry]
name = "nexus-backend"
version = "0.1.0"
description = ""
authors = ["j1ere <jeremywambua106@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "<5.2"
djangorestframework = "^3.16.0"
pillow = "^11.2.1"
python-dotenv = "^1.1.0"
django-cors-headers = "^4.7.0"
psycopg2 = "^2.9.10"
djangorestframework-simplejwt = "^5.5.0"
daphne = "^4.2.0"
channels = "^4.2.2"
channels-redis = "^4.2.1"
django-extensions = "^4.1"
django-filter = "^25.1"
psycopg2-binary = "^2.9.10"
dj-database-url = "^3.0.1"
python-decouple = "^3.8"


[tool.poetry.group.dev.dependencies]
uvicorn = {extras = ["standard"], version = "^0.34.3"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
